requirements.txt
requests_oauthlib
gistyc
oauthlib
PyYAML
xlrd
urllib3
pytz
pathvalidate
textblob
bs4